# GGRC-5028 Process CAD getter only for CustomAttributable objects